backend/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "flooring-crm-backend"
version = "1.0.0"
description = "Professional Flooring Company Management System Backend"
authors = ["Flooring CRM Development Team <support@flooringcrm.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.109.2"}
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
python-multipart = "^0.0.9"
email-validator = "^2.1.0.post1"
httpx = "^0.26.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
bcrypt = "^4.1.2"
asyncpg = "^0.29.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
ruff = "^0.1.6"

[tool.poetry.urls]
repository = "https://github.com/flooring-crm/backend"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=app --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []
